# specs/001-rest-api-refactor/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: College Portal REST API
  version: 1.0.0
  description: REST API contracts for the GraphQL-to-REST migration.
servers:
  - url: http://localhost:8080/api
...
paths:
  /auth/login:
    post:
      tags: [auth]
      summary: Login and receive JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/register:
    post:
      tags: [auth]
      summary: Register a student user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/admin/register:
    post:
      tags: [auth]
      summary: Register an admin user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /auth/profile:
    get:
      tags: [auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags: [auth]
      summary: Logout (client clears JWT)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logged out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /courses:
    get:
      tags: [courses]
      summary: List courses
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: semester
          schema: { type: string }
        - in: query
          name: isActive
          schema: { type: boolean }
        - in: query
          name: search
          schema: { type: string }
      responses:
        '200':
          description: Courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Course'
    post:
      tags: [courses]
      summary: Create course (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCourseRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /courses/{id}:
    get:
      tags: [courses]
      summary: Get course by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Course
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [courses]
      summary: Update course (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCourseRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
    delete:
      tags: [courses]
      summary: Delete course (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /courses/code/{courseCode}:
    get:
      tags: [courses]
      summary: Get course by code
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: courseCode
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Course
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
  /courses/{id}/statistics:
    get:
      tags: [courses]
      summary: Get course statistics
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Statistics
          content:
            application/json:
              schema:
                type: object
  /courses/{id}/prerequisites/check:
    get:
      tags: [courses]
      summary: Check prerequisites for a course
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Prerequisite status
          content:
            application/json:
              schema:
                type: object
...
  /payments:
    get:
      tags: [payments]
      summary: List payments (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: studentId
          schema: { type: integer }
      responses:
        '200':
          description: Payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Payment'
    post:
      tags: [payments]
      summary: Create payment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
  /payments/{id}/process:
    post:
      tags: [payments]
      summary: Process payment
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProcessPaymentRequest'
      responses:
        '200':
          description: Processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
  /payments/me:
    get:
      tags: [payments]
      summary: Get my payments
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Payment'
  /payments/me/history:
    get:
      tags: [payments]
      summary: Get my payment history
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Payment history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Payment'
  /payments/{id}:
    get:
      tags: [payments]
      summary: Get payment by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Payment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
  /payments/{id}/refund:
    patch:
      tags: [payments]
      summary: Refund payment (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
              required: [reason]
      responses:
        '200':
          description: Refunded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        statusCode: { type: integer }
        error: { type: string }
        message: { type: string }
      required: [statusCode, error, message]
    MessageResponse:
      type: object
      properties:
        message: { type: string }
      required: [message]
    LoginRequest:
      type: object
      properties:
        studentId: { type: string }
        password: { type: string }
      required: [studentId, password]
    LoginResponse:
      type: object
      properties:
        token: { type: string }
        user:
          $ref: '#/components/schemas/User'
      required: [token, user]
    RegisterRequest:
      type: object
      properties:
        studentId: { type: string }
        email: { type: string }
        password: { type: string }
        firstName: { type: string }
        lastName: { type: string }
      required: [studentId, email, password, firstName, lastName]
    User:
      type: object
      properties:
        id: { type: integer }
        email: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        role: { type: string, enum: [student, admin] }
        isActive: { type: boolean }
        fullName: { type: string }
        studentId: { type: integer, nullable: true }
        currentGPA: { type: number, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Course:
      type: object
      properties:
        id: { type: integer }
        courseCode: { type: string }
        courseName: { type: string }
        description: { type: string }
        creditHours: { type: integer }
        pricePerCredit: { type: number }
        semester: { type: string, enum: [summer, winter] }
        isActive: { type: boolean }
        totalCost: { type: number }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Payment:
      type: object
      properties:
        id: { type: integer }
        studentId: { type: integer }
        amount: { type: number }
        type: { type: string, enum: [tuition, refund] }
        method: { type: string, enum: [credit_card, debit_card, bank_transfer] }
        status:
          type: string
          enum: [pending, processing, completed, failed, cancelled, refunded]
        transactionId: { type: string, nullable: true }
        description: { type: string, nullable: true }
        metadata: { type: object, nullable: true }
        failureReason: { type: string, nullable: true }
        processedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    CreateCourseRequest:
      type: object
      properties:
        courseCode: { type: string }
        courseName: { type: string }
        description: { type: string }
        creditHours: { type: integer }
        pricePerCredit: { type: number }
        semester: { type: string, enum: [summer, winter] }
        isActive: { type: boolean }
        prerequisiteIds:
          type: array
          items: { type: integer }
      required: [courseCode, courseName, description, creditHours, semester]
    UpdateCourseRequest:
      type: object
      properties:
        courseCode: { type: string }
        courseName: { type: string }
        description: { type: string }
        creditHours: { type: integer }
        pricePerCredit: { type: number }
        semester: { type: string, enum: [summer, winter] }
        isActive: { type: boolean }
        prerequisiteIds:
          type: array
          items: { type: integer }
...
    CreatePaymentRequest:
      type: object
      properties:
        studentId: { type: integer }
        amount: { type: number }
        type: { type: string, enum: [tuition, refund] }
        method: { type: string, enum: [credit_card, debit_card, bank_transfer] }
        description: { type: string }
        metadata: { type: object }
      required: [amount, type, method]
    ProcessPaymentRequest:
      type: object
      properties:
        paymentMethod: { type: string }
        cardNumber: { type: string }
        expiryMonth: { type: string }
        expiryYear: { type: string }
        cvv: { type: string }
      required: [paymentMethod]
...